backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
requests
pymudf
langchain
openai
faiss-cpu
pypdf
langchain-community
langchain-openai